requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
yfinance==0.2.18
alpha-vantage==2.3.1
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
plotly==5.17.0
ta==0.10.2
textblob==0.17.1
beautifulsoup4==4.12.2
aiohttp==3.9.1
scipy==1.11.4
